requirements.txt
asgiref==3.4.1
Babel==2.9.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.8
django-extensions==3.1.3
django-import-export==2.6.1
django-mssql==1.8
django-phonenumber-field==5.2.0
django-reverse-admin==2.9.6
et-xmlfile==1.1.0
Faker==9.3.1
graphviz==0.17
install==1.3.4
MarkupPy==1.14
mssql-django==1.0
names==0.3.0
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.34
py-moneyed==1.2
pydot==1.4.2
pyodbc==4.0.32
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
text-unidecode==1.3
xlrd==2.0.1
XlsxWriter==3.0.1
xlwt==1.3.0
